ai_cmaker/pyproject.toml
[project]
name = "ai-cmaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioboto3==12.2.0",
    "aiobotocore==2.9.0",
    "aiogram>=3.19.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "arq>=0.26.3",
    "boto3==1.33.13",
    "botocore==1.33.13",
    "ffmpeg-python>=0.2.0",
    "greenlet>=3.1.1",
    "httpx>=0.28.1",
    "openai>=1.69.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.40",
]
